requirements.txt
Django<2.0.0,>=1.11.8
mysql-python==1.2.5
django-enumfields==0.10.1
